% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gracefully_fail}
\alias{gracefully_fail}
\title{Fail gracefully if Wikisource resources are not available}
\usage{
gracefully_fail(remote_url)
}
\arguments{
\item{remote_url}{A remote URL.}
}
\description{
The function allows to fail gracefully with an informative message 
if the Wikisource resource is not available (and not give a check warning nor error).
}
\details{
See full discussion to be compliante with the CRAN policy 
<https://community.rstudio.com/t/internet-resources-should-fail-gracefully/49199>
}
