% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.wikisource}
\alias{print.wikisource}
\title{Print method for Wikisource objects}
\usage{
\method{print}{wikisource}(x, ...)
}
\arguments{
\item{x}{wikisource object from get_item, get_random_item, get_property or get_random_property}

\item{\dots}{Arguments to be passed to methods}
}
\description{
print found objects generally.
}
\seealso{
get_item, get_random_item, get_property or get_random_property
}

