% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikisourceR.R
\docType{package}
\name{WikisourceR}
\alias{WikisourceR}
\alias{WikisourceR-package}
\title{API client library for Wikisource}
\description{
This package serves as an API client for \href{Wikisource}{https://www.wikisource.org}.
See the accompanying vignette for more details.
}
\seealso{
\code{\link{get_random}} for selecting a random item or property,
\code{\link{get_item}} for a /specific/ item or property, or \code{\link{find_item}}
for using search functionality to pull out item or property IDs where the descriptions
or aliases match a particular search term.
}

